seal SparkeyUri trait
