--- scio-extra/src/main/scala/com/spotify/scio/extra/sparkey/instances/CoderInstances.scala
@@ -1,29 +1,0 @@
-/*
- * Copyright 2023 Spotify AB.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package com.spotify.scio.extra.sparkey.instances
-
-import com.spotify.scio.coders.Coder
-import com.spotify.scio.extra.sparkey.{ShardedSparkeyUri, SparkeyUri}
-
-trait CoderInstances {
-  implicit def sparkeyUriCoder: Coder[SparkeyUri] =
-    Coder.xmap[String, SparkeyUri](Coder[String])(SparkeyUri(_), _.basePath)
-
-  implicit def shardedSparkeyUriCoder: Coder[ShardedSparkeyUri] =
-    Coder.xmap[String, ShardedSparkeyUri](Coder[String])(ShardedSparkeyUri(_), _.basePath)
-}
