+++ scio-extra/src/main/scala/com/spotify/scio/extra/sparkey/package.scala
@@ -120,7 +120,7 @@ import scala.util.hashing.MurmurHash3
  *   .toSCollection
  * }}}
  */
-package object sparkey extends SparkeyReaderInstances with CoderInstances {
+package object sparkey extends SparkeyReaderInstances {
 
   /** Enhanced version of [[ScioContext]] with Sparkey methods. */
   implicit class SparkeyScioContext(private val self: ScioContext) extends AnyVal {
